+++ packages/throws_collector/lib/throws_collector.dart
@@ -4,6 +4,8 @@ export 'src/options.dart'
     show
         CollectorConfig,
         CollectorInput,
+        DartSdkInputSource,
+        FlutterSdkInputSource,
         GitInputSource,
         InputSource,
         LocalInputSource,